refactor (LottoTickets.java) : 자동/수동 생성기능을 팩토리 메서드로 구현
